// src/models/Farmer.model.ts
export interface FarmerModel {
    farmer_id: string;
    title_en: string;
    title_ar: string;
    description_en: string;
    description_ar: string;
    active: string;
    quality_rate: string;
    delivery_rate: string;
    final_rate: string;
    value_rate: string;
    image: string;
    lat: string;
    lng: string;
    location: string;
}